Separate anchored and unanchored annotations into tabs.
Default to the annotations tab when user switches between groups, clears a selection or reloads a page.
The ANNOTATIONS_SYNCED event is emitted when annotations for a document have finished anchoring. Because annotations are not marked as orphans until they fail to anchor, select the annotations tab after they have finished anchoring.

Test cases:
-------------
On page load:
Annotations tab is selected.

On group change when annotations tab was selected for the previous group:
Annotations tab is selected.

On group change when the notes/orphans tab was selected for the previous group:
Annotations tab is selected.

On group change to a group which has no notes/annotations/orphans when a notes/orphans tab is selected:
Annotations tab is selected.

On group change to a group which has no notes/annotations/orphans when an annotations tab is selected:
Annotations tab is selected.

After annotations are anchored:
Annotations tab is selected.

When an annotation/note/orphan is selected:
Select annotations tab or tabTypeForSelection to display the annotation/note/orphan.
Then select the annotations/notes tab when the selection is cleared.

https://trello.com/c/1NVK2jwv/400-add-a-tab-for-unanchored-annotations
https://trello.com/c/OLdLTlLT/342-separate-annotations-and-notes

Hide orphans tab behind orphans_tab flag.

Add tests.