Fix direct links to Page Notes
The logic to change the selected tab when a direct-linked Page Note
was loaded was never triggered because page notes are never anchored
so no ANNOTATIONS_SYNCED event is triggered.

Instead update the selected tab when an annotation is loaded if the
annotation is already selected, or when the anchoring status of a
selected annotation changes.

 * Add logic to ADD_ANNOTATIONS and UPDATE_ANCHR_STATUS action
   handlers to update the active tab if an already-selected annotation
   is loaded or a selected annotation moves to a different tab

 * Update fixtures for saved annotations to include a quote
   selector so that `metadata.isAnnotation()` returns true.

   This required enabling ngSanitize for the <annotation> component test
   because the content of the quote selector is used together with
   "ng-bind-html"