Retry anchoring if the range no longer matches the annotation when we come to highlight the text
The process of highlighting a range for one annotation can invalidate
the anchored Range for another annotation. This typically occurs when
the original annotations overlap.

Work around this issue by testing whether the anchored Range's content
matches the annotations quote before inserting the highlight. If it does
not, retry anchoring.