Fix handling of persisted OAuth tokens for anonymous third-party users
When automatic login to a third-party account is being used, via a grant
token, OAuth tokens persisted by other client sessions should not be
used.

Fix this in the case where the user is anonymous on the publisher's
website and so the publisher will have set the "grantToken" property in
the "services" array to `null`.

Also add a test for the case where the user is logged in and the grant
token is a JWT (although opaque to the client). This case was already
working.