Replace "usernameLink" config option with callback
Give publishers more flexibility in how they handle username links in
annotation cards being clicked by replacing the "usernameLink" config
option with an `onUserLinkClicked` callback in the service config.

This addresses feedback from eLife that they wanted the username
link not to open in a new window. It is also more consistent with how
the publisher handles other links in the sidebar such as Help, the
user's profile and Sign Up / Log In.