Convert cross-origin-rpc to a service which uses util/postmessage-json-rpc
Convert the cross-origin-rpc utility module to a service following the
conventions of other services in the sidebar app and replace its
internal RPC handling with use of the RPC server class defined in the
util/postmessage-json-rpc module.

This means that all details of cross-frame RPCs are now handled in one
utility module and concerns are better separated.