Use main/canonical document URL when searching for groups
Use the same URL that will be used as the "main URL" [1] for annotations
created on the current page, when searching for groups associated with
the current page.

This fixes an issue where a page could declare additional URLs in its metadata,
and one of those URLs (collectively, the "search URIs") could end up being used
to search for groups, instead of the "main" URL. In the scenario where there are
multiple frames connected, always use the URL of the "main" frame to search for
groups.

 - Add a `mainFrame` selector to the store which returns the main/top-level
   frame out of the frames that are connected
 - Instead of taking the first HTTP(S) URL from `store.searchUris` as the URL
   to use for fetching associated groups in sidebar/services/groups.js, use
   `store.mainFrame().uri` instead, which returns the main URL of the main
   frame. The `uri` property of a frame should always match the URI that is
   used as the primary URI of annotations created in a particular frame.

[1] Frames that are connected to the sidebar may have multiple URIs
    associated with them, obtained from sources including browser APIs
    (`window.location.href`) and metadata present on the page.

    One of these is designated the "primary" URL. For HTML documents, this is
    the URL specified by `<link rel="canonical">` if present or
    `window.location.href` otherwise. For PDF documents, this is the URL
    returned by PDF.js as the source of the document, or the PDF fingerprint
    otherwise.